fix(API): move alpha update plugin to internal
The Alpha Update plugin was exposed in the public API. Considering it is meant for CLI users only, its API is not well-documented and is not stable, it should belong in internal.
