+++ internal/plugins/optional/autoupdate/v1alpha/edit_test.go
