+++ internal/plugins/optional/autoupdate/v1alpha/init.go
@@ -22,52 +22,41 @@ import (
 
 	"github.com/spf13/pflag"
 
+	"sigs.k8s.io/kubebuilder/v4/internal/plugins/optional/autoupdate/v1alpha/scaffolds"
 	"sigs.k8s.io/kubebuilder/v4/pkg/config"
 	"sigs.k8s.io/kubebuilder/v4/pkg/machinery"
 	"sigs.k8s.io/kubebuilder/v4/pkg/plugin"
-	"sigs.k8s.io/kubebuilder/v4/pkg/plugins/optional/autoupdate/v1alpha/scaffolds"
 )
 
-var _ plugin.EditSubcommand = &editSubcommand{}
+var _ plugin.InitSubcommand = &initSubcommand{}
 
-type editSubcommand struct {
+type initSubcommand struct {
 	config      config.Config
 	useGHModels bool
 }
 
-func (p *editSubcommand) UpdateMetadata(cliMeta plugin.CLIMetadata, subcmdMeta *plugin.SubcommandMetadata) {
+func (p *initSubcommand) UpdateMetadata(cliMeta plugin.CLIMetadata, subcmdMeta *plugin.SubcommandMetadata) {
 	subcmdMeta.Description = metaDataDescription
 
-	subcmdMeta.Examples = fmt.Sprintf(`  # Edit a common project with this plugin
-  %[1]s edit --plugins=%[2]s
+	subcmdMeta.Examples = fmt.Sprintf(`  # Initialize a common project with this plugin
+  %[1]s init --plugins=%[2]s
 
-  # Edit a common project with GitHub Models enabled (requires repo permissions)
-  %[1]s edit --plugins=%[2]s --use-gh-models
+  # Initialize with GitHub Models enabled (requires repo permissions)
+  %[1]s init --plugins=%[2]s --use-gh-models
 `, cliMeta.CommandName, plugin.KeyFor(Plugin{}))
 }
 
-func (p *editSubcommand) BindFlags(fs *pflag.FlagSet) {
+func (p *initSubcommand) BindFlags(fs *pflag.FlagSet) {
 	fs.BoolVar(&p.useGHModels, "use-gh-models", false,
 		"enable GitHub Models AI summary in the scaffolded workflow (requires GitHub Models permissions)")
 }
 
-func (p *editSubcommand) InjectConfig(c config.Config) error {
+func (p *initSubcommand) InjectConfig(c config.Config) error {
 	p.config = c
 	return nil
 }
 
-func (p *editSubcommand) PreScaffold(machinery.Filesystem) error {
-	if len(p.config.GetCliVersion()) == 0 {
-		return fmt.Errorf(
-			"you must manually upgrade your project to a version that records the CLI version in PROJECT (`cliVersion`) " +
-				"to allow the `alpha update` command to work properly before using this plugin.\n" +
-				"More info: https://book.kubebuilder.io/migrations",
-		)
-	}
-	return nil
-}
-
-func (p *editSubcommand) Scaffold(fs machinery.Filesystem) error {
+func (p *initSubcommand) Scaffold(fs machinery.Filesystem) error {
 	if err := insertPluginMetaToConfig(p.config, PluginConfig{UseGHModels: p.useGHModels}); err != nil {
 		return fmt.Errorf("error inserting project plugin meta to configuration: %w", err)
 	}
@@ -75,13 +64,13 @@ func (p *editSubcommand) Scaffold(fs machinery.Filesystem) error {
 	scaffolder := scaffolds.NewInitScaffolder(p.useGHModels)
 	scaffolder.InjectFS(fs)
 	if err := scaffolder.Scaffold(); err != nil {
-		return fmt.Errorf("error scaffolding edit subcommand: %w", err)
+		return fmt.Errorf("error scaffolding init subcommand: %w", err)
 	}
 
 	return nil
 }
 
-func (p *editSubcommand) PostScaffold() error {
+func (p *initSubcommand) PostScaffold() error {
 	// Inform users about GitHub Models if they didn't enable it
 	if !p.useGHModels {
 		log.Info("Consider enabling GitHub Models to get an AI summary to help with the update")
