+++ internal/plugins/optional/autoupdate/v1alpha/plugin.go
