+++ internal/plugins/optional/autoupdate/v1alpha/plugin_test.go
