+++ internal/plugins/optional/autoupdate/v1alpha/scaffolds/internal/github/auto_update.go
