+++ internal/plugins/optional/autoupdate/v1alpha/suite_test.go
